--- types/react-keyboard-event-handler/KeyboardEventHandler.d.ts
@@ -1,14 +1,0 @@
-import * as React from "react";
-
-export interface KeyboardEventHandlerProps {
-    handleKeys?: readonly string[];
-    handleEventType?: "keydown" | "keyup" | "keypress";
-    handleFocusableElements?: boolean;
-    onKeyEvent?: (key: string, event: KeyboardEvent) => void;
-    isDisabled?: boolean;
-    isExclusive?: boolean;
-    children?: React.ReactNode;
-}
-
-declare const KeyboardEventHandler: React.SFC<KeyboardEventHandlerProps>;
-export default KeyboardEventHandler;
