+++ types/react-keyboard-event-handler/index.d.ts
@@ -3,4 +3,17 @@
 // Definitions by: Jamie Gerrard Lievesley <https://github.com/jamiegluk>
 // Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
 
-export { default } from "./KeyboardEventHandler";
+import * as React from "react";
+
+export interface KeyboardEventHandlerProps {
+    handleKeys?: readonly string[];
+    handleEventType?: "keydown" | "keyup" | "keypress";
+    handleFocusableElements?: boolean;
+    onKeyEvent?: (key: string, event: KeyboardEvent) => void;
+    isDisabled?: boolean;
+    isExclusive?: boolean;
+    children?: React.ReactNode;
+}
+
+declare const KeyboardEventHandler: React.SFC<KeyboardEventHandlerProps>;
+export default KeyboardEventHandler;
